app/requirements.txt
flask==2.1.2
python-dotenv
requests
html2text
pytest
werkzeug==2.1.2
